% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fer_dressage.R
\name{add_category_n}
\alias{add_category_n}
\title{Determine level for N champioship}
\usage{
add_category_n(df, x, nb_cat, max_pt)
}
\arguments{
\item{df}{A results dataframe}

\item{x}{Number of categories to consider}

\item{nb_cat}{Max number of S results}

\item{max_pt}{Max number of points for M level}
}
\value{
A results dataframe
}
\description{
Determine level for N champioship
}
