% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_dressage.R
\name{add_category_r}
\alias{add_category_r}
\title{Determine level for R champioship}
\usage{
add_category_r(df, x, nb_cat, max_pt)
}
\arguments{
\item{df}{A results dataframe}

\item{x}{Number of categories to consider}

\item{nb_cat}{Max number of M results}

\item{max_pt}{Max number of points for L level}
}
\value{
A results dataframe
}
\description{
Determine level for R champioship
}
