% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fer_dressage.R
\name{add_category_u21}
\alias{add_category_u21}
\title{Determine level for U21 champioship}
\usage{
add_category_u21(df)
}
\arguments{
\item{df}{A results dataframe}
}
\value{
A results dataframe
}
\description{
Determine level for U21 champioship
}
