% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_riders.R
\name{add_rider_infos}
\alias{add_rider_infos}
\title{Add rider info to result dataframe}
\usage{
add_rider_infos(df)
}
\arguments{
\item{df}{A dataframe produced by `map_fnch_class_results`}
}
\value{
A dataframe
}
\description{
Add rider info to result dataframe
}
