% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aen_jumping.R
\name{aen_check_championship_br}
\alias{aen_check_championship_br}
\title{Check qualifications for B/R Championship}
\usage{
aen_check_championship_br(table)
}
\arguments{
\item{table}{The data table}
}
\value{
The data table with a checkout column
}
\description{
Check qualifications for B/R Championship
}
