% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aen_jumping.R
\name{aen_check_championship_rn}
\alias{aen_check_championship_rn}
\title{Check qualifications for R/N Championship}
\usage{
aen_check_championship_rn(table)
}
\arguments{
\item{table}{The data table}
}
\value{
The data table with a checkout column
}
\description{
Check qualifications for R/N Championship
}
