% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aen_jumping.R
\name{aen_check_rn_championship}
\alias{aen_check_rn_championship}
\title{Check qualification for RN Championship}
\usage{
aen_check_rn_championship(participant, last_championship)
}
\arguments{
\item{participant}{A list with riderid and horseid}

\item{last_championship}{The date of the last championship}
}
\value{
A list of data points with qualification decision
}
\description{
Check qualification for RN Championship
}
