% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aen_jumping.R
\name{aen_get_championship_data}
\alias{aen_get_championship_data}
\title{Add data to startlist}
\usage{
aen_get_championship_data(table, last_championship, engagement_deadline)
}
\arguments{
\item{table}{The startlist}

\item{last_championship}{Date of the last championship}

\item{engagement_deadline}{Deadline of current championship}
}
\value{
The table with data
}
\description{
Add data to startlist
}
