% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aen_jumping.R
\name{aen_get_championship_infos}
\alias{aen_get_championship_infos}
\title{Get all the data needed for Championship checks}
\usage{
aen_get_championship_infos(
  rider_id,
  horse_id,
  last_championship,
  engagement_deadline
)
}
\arguments{
\item{rider_id}{The rider id}

\item{horse_id}{The horse id}

\item{last_championship}{Date of the last championship}

\item{engagement_deadline}{Deadline of current championship}
}
\value{
A list to be expanded in championship table
}
\description{
Get all the data needed for Championship checks
}
