% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{apply_event_colors}
\alias{apply_event_colors}
\title{Apply style colors}
\usage{
apply_event_colors(wb, sheet, ...)
}
\arguments{
\item{wb}{The workbook to write in}

\item{sheet}{The sheet to write in}

\item{...}{The current row values}
}
\description{
Apply style colors
}
