% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{browse_events_fnch}
\alias{browse_events_fnch}
\title{Browse events.swiss-equestrian.ch}
\usage{
browse_events_fnch(url, query = NULL)
}
\arguments{
\item{url}{A string}

\item{query}{A list of query arguments}
}
\value{
An rvest response object
}
\description{
Browse events.swiss-equestrian.ch
}
