% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{browse_my_fnch}
\alias{browse_my_fnch}
\title{Browse my.fnch.ch}
\usage{
browse_my_fnch(url, query = NULL)
}
\arguments{
\item{url}{A string}

\item{query}{A list of query arguments}
}
\value{
An rvest response object
}
\description{
Browse my.fnch.ch
}
