% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_dressage.R
\name{calculate_judges_mean}
\alias{calculate_judges_mean}
\title{Calculate mean grade from judges' note}
\usage{
calculate_judges_mean(obj)
}
\arguments{
\item{obj}{A result object}
}
\value{
A numeric mean grade
}
\description{
Calculate mean grade from judges' note
}
