% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{check_calendar_names}
\alias{check_calendar_names}
\title{Translate columns into french}
\usage{
check_calendar_names(df)
}
\arguments{
\item{df}{A calendar dataframe}
}
\value{
A cleaned up calendar dataframe
}
\description{
Translate columns into french
}
\keyword{internal}
