% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{create_collision_table}
\alias{create_collision_table}
\title{Create collision table}
\usage{
create_collision_table(calendar, ...)
}
\arguments{
\item{...}{The current row values}
}
\description{
Create collision table
}
