% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{create_ics_cal}
\alias{create_ics_cal}
\title{Create ICS Event for}
\usage{
create_ics_cal(...)
}
\arguments{
\item{...}{Must contain Start, End, Summary, Description and URL}
}
\value{
A calendar ICS event
}
\description{
Create ICS Event for
}
