% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_dr_results.R
\name{event_canton_flag}
\alias{event_canton_flag}
\title{Return flag url}
\usage{
event_canton_flag(id)
}
\arguments{
\item{id}{Canton short id}
}
\value{
A string url
}
\description{
Return flag url
}
\keyword{internal}
