% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_dr_results.R
\name{event_dr_read_result_file}
\alias{event_dr_read_result_file}
\title{Read the Rosson dressage results}
\usage{
event_dr_read_result_file(path)
}
\arguments{
\item{path}{A file path}
}
\value{
A dataframe
}
\description{
Reads a Rosson result file for a dressage class
and renames judge columns to just the letter.
}
\keyword{internal}
