% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_dr_results.R
\name{event_dr_results}
\alias{event_dr_results}
\title{Results table for dressage classes}
\usage{
event_dr_results(file, provisional = FALSE)
}
\arguments{
\item{file}{The results file}

\item{provisional}{Boolean: is the result provisional?}
}
\value{
A reactable object
}
\description{
Results table for dressage classes
}
