% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_dr_results.R
\name{event_dr_results_intercantonal}
\alias{event_dr_results_intercantonal}
\title{Results table for dressage classes with canton flag}
\usage{
event_dr_results_intercantonal(file, affiliation, provisional = FALSE)
}
\arguments{
\item{file}{The results file}

\item{affiliation}{File with the canton affiliations}

\item{provisional}{Boolean: is the result provisional?}
}
\value{
A reactable object
}
\description{
Results table for dressage classes with canton flag
}
