% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_dr_results.R
\name{event_dr_results_intercantonal_final}
\alias{event_dr_results_intercantonal_final}
\title{Results table for dressage classes with canton flag}
\usage{
event_dr_results_intercantonal_final(
  file_round1,
  file_round2 = "",
  affiliation
)
}
\arguments{
\item{file_round1}{The result file of the first round}

\item{file_round2}{The result file of the second round}

\item{affiliation}{File with the canton affiliations}
}
\value{
A reactable object
}
\description{
Results table for dressage classes with canton flag
}
