% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_dr_results.R
\name{event_dr_results_intercantonal_individual}
\alias{event_dr_results_intercantonal_individual}
\title{Results table for dressage intercantonal individual}
\usage{
event_dr_results_intercantonal_individual(
  file_round1,
  file_round2,
  affiliation,
  provisional = FALSE
)
}
\arguments{
\item{file_round1}{The result file of the first round}

\item{file_round2}{The result file of the second round}

\item{affiliation}{File with the canton affiliations}

\item{provisional}{Boolean: is the result provisional?}
}
\value{
A reactable object
}
\description{
Results table for dressage intercantonal individual
}
