% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_dr_results.R
\name{event_dr_results_table}
\alias{event_dr_results_table}
\title{Format the results in a reactable table}
\usage{
event_dr_results_table(
  results,
  last_rider,
  championship = FALSE,
  intercantonal = FALSE,
  individual = FALSE
)
}
\arguments{
\item{results}{The results dataframe}

\item{last_rider}{The last rider (if highlighting is desired)}

\item{championship}{Boolean for championship tables}
}
\value{
A reactable object
}
\description{
Format the results in a reactable table
}
\keyword{internal}
