% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_accounting}
\alias{fnch_accounting}
\title{FSSE Accounting: Sum up and format}
\usage{
fnch_accounting(tbl)
}
\arguments{
\item{tbl}{An FSSE Accounting tibble}
}
\value{
A flextable
}
\description{
FSSE Accounting: Sum up and format
}
