% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_add_judge}
\alias{fnch_add_judge}
\title{FSSE Accounting: Add row to judges table}
\usage{
fnch_add_judge(df, name, classes, half_days, days)
}
\arguments{
\item{df}{A judge table}

\item{name}{Name of judge}

\item{half_days}{Number of half days judged}

\item{days}{Number of days judged}
}
\value{
A tibble
}
\description{
FSSE Accounting: Add row to judges table
}
