% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_logbook.R
\name{fnch_add_logbook_event}
\alias{fnch_add_logbook_event}
\title{Insert a log event into notion}
\usage{
fnch_add_logbook_event(
  event_name,
  event_judge,
  event_horse_id,
  event_horse,
  event_rider_id,
  event_rider,
  event_categories,
  event_comment
)
}
\arguments{
\item{event_name}{The event location}

\item{event_judge}{The judge id}

\item{event_horse_id}{A numeric passeport id}

\item{event_horse}{The Horse name}

\item{event_rider_id}{A numeric license id}

\item{event_rider}{The rider name}

\item{event_categories}{A vector of categories}

\item{event_comment}{A comment}
}
\description{
Insert a log event into notion
}
