% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_add_old_judge}
\alias{fnch_add_old_judge}
\title{FSSE Accounting: Add row to judges table (old system)}
\usage{
fnch_add_old_judge(df, name, classes)
}
\arguments{
\item{df}{A judges table (old system)}

\item{name}{Name of judge}

\item{classes}{Number of classes}
}
\value{
A tibble
}
\description{
FSSE Accounting: Add row to judges table (old system)
}
