% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_add_presidency}
\alias{fnch_add_presidency}
\title{FSSE Accounting: Add row to presidency table}
\usage{
fnch_add_presidency(df, name, days)
}
\arguments{
\item{df}{A presidency table}

\item{name}{Name of president}

\item{days}{Number of days}
}
\value{
A tibble
}
\description{
FSSE Accounting: Add row to presidency table
}
