% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_candidates_base}
\alias{fnch_candidates_base}
\title{FSSE Accounting: Base candidate table}
\usage{
fnch_candidates_base()
}
\value{
A tibble
}
\description{
FSSE Accounting: Base candidate table
}
