% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_markdown.R
\name{fnch_document}
\alias{fnch_document}
\title{Standard pdf document}
\usage{
fnch_document(..., metadata = c())
}
\arguments{
\item{metadata}{Additional pandoc metadata}
}
\description{
Standard pdf document
}
