% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fnch_file}
\alias{fnch_file}
\title{A shortcut function for the system.file to the rsvps package}
\usage{
fnch_file(...)
}
\value{
A path
}
\description{
A shortcut function for the system.file to the rsvps package
}
