% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_get_class_data}
\alias{fnch_get_class_data}
\title{Get data for a specific class}
\usage{
fnch_get_class_data(df, class_id)
}
\arguments{
\item{df}{Jury data table}

\item{class_id}{Id of the class}
}
\value{
A list
}
\description{
Get data for a specific class
}
\keyword{internal}
