% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_get_class_headers}
\alias{fnch_get_class_headers}
\title{Get data for a specific class}
\usage{
fnch_get_class_headers(class_id, df)
}
\arguments{
\item{class_id}{Id of the class}

\item{df}{Jury data table}
}
\value{
A datafrane
}
\description{
Get data for a specific class
}
