% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_get_judge_cols}
\alias{fnch_get_judge_cols}
\title{Return judge data}
\usage{
fnch_get_judge_cols(judge_id, df)
}
\arguments{
\item{judge_id}{The judge id}

\item{df}{The judge data table}
}
\value{
A tibble
}
\description{
Return judge data
}
