% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_get_judge_data}
\alias{fnch_get_judge_data}
\title{Get data for a specific judge}
\usage{
fnch_get_judge_data(df, judge_id)
}
\arguments{
\item{df}{Judge data table}

\item{judge_id}{Id of the judge}
}
\value{
A list
}
\description{
Get data for a specific judge
}
\keyword{internal}
