% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_get_judge_function_abbr}
\alias{fnch_get_judge_function_abbr}
\title{Get judge role abbreviation}
\usage{
fnch_get_judge_function_abbr(role)
}
\arguments{
\item{role}{Judge role}
}
\description{
Get judge role abbreviation
}
\keyword{internal}
