% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_get_judge_functions}
\alias{fnch_get_judge_functions}
\title{Get judge functions}
\usage{
fnch_get_judge_functions()
}
\value{
A vector of judge roles
}
\description{
Get judge functions
}
\keyword{internal}
