% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_get_judge_table}
\alias{fnch_get_judge_table}
\title{Get jury table for a specific date}
\usage{
fnch_get_judge_table(df, filter_date)
}
\arguments{
\item{df}{Jury data table}

\item{filter_date}{Date to filter for}
}
\value{
A tibble
}
\description{
Get jury table for a specific date
}
