% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_get_jury_data}
\alias{fnch_get_jury_data}
\title{Get FSSE Jury for given tag and dates}
\usage{
fnch_get_jury_data(tag, first_day, last_day)
}
\arguments{
\item{tag}{An event tag}

\item{first_day}{Date of first class}

\item{last_day}{Date of last class}
}
\value{
A tibble
}
\description{
Get FSSE Jury for given tag and dates
}
\keyword{internal}
