% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_get_jury_dates}
\alias{fnch_get_jury_dates}
\title{Get the table with date and responsible jury presidents}
\usage{
fnch_get_jury_dates(jury_data, judges)
}
\arguments{
\item{jury_data}{A tibble with classes data}

\item{judges}{A tibble with judges data}
}
\value{
A tibble
}
\description{
Get the table with date and responsible jury presidents
}
\keyword{internal}
