% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_knit_judge_table}
\alias{fnch_knit_judge_table}
\title{Knit jury for one day}
\usage{
fnch_knit_judge_table(date, nom, df)
}
\arguments{
\item{date}{Date}

\item{nom}{Name of the responsible jury president}

\item{df}{Jury data table}
}
\value{
A knitted string
}
\description{
Knit jury for one day
}
