% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_markdown.R
\name{fnch_letter_standard}
\alias{fnch_letter_standard}
\title{Standard pdf letter}
\usage{
fnch_letter_standard(..., metadata = c())
}
\arguments{
\item{metadata}{Additional pandoc metadata}
}
\description{
Standard pdf letter
}
