% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_main_courante}
\alias{fnch_main_courante}
\title{FSSE Jury: event table}
\usage{
fnch_main_courante(nrow = 20)
}
\arguments{
\item{nrow}{Number of rows}
}
\value{
A flextable
}
\description{
FSSE Jury: event table
}
