% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_map_judge_ids}
\alias{fnch_map_judge_ids}
\title{Get judge ids}
\usage{
fnch_map_judge_ids(data)
}
\arguments{
\item{data}{A list of notion objects}
}
\value{
A list of ids
}
\description{
Get judge ids
}
\keyword{internal}
