% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_pdftable_jury}
\alias{fnch_pdftable_jury}
\title{Get PDF Jury tables}
\usage{
fnch_pdftable_jury(tag, first_day, last_day)
}
\arguments{
\item{tag}{An event tag}

\item{first_day}{Date of first class}

\item{last_day}{Date of last class}
}
\value{
Knitr markdown
}
\description{
Get PDF Jury tables
}
