% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_presidency_base}
\alias{fnch_presidency_base}
\title{FSSE Accounting: Base presidency table}
\usage{
fnch_presidency_base()
}
\value{
A tibble
}
\description{
FSSE Accounting: Base presidency table
}
