% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_react_sticky_style}
\alias{fnch_react_sticky_style}
\title{Helper function for sticky columns}
\usage{
fnch_react_sticky_style(left = TRUE)
}
\arguments{
\item{left}{Boolean: align left?}
}
\description{
Helper function for sticky columns
}
\keyword{internal}
