% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_reactable_jury}
\alias{fnch_reactable_jury}
\title{Get reactable Jury table}
\usage{
fnch_reactable_jury(tag, first_day, last_day)
}
\arguments{
\item{tag}{An event tag}

\item{first_day}{Date of first class}

\item{last_day}{Date of last class}
}
\value{
A reactable
}
\description{
Get reactable Jury table
}
