% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_markdown.R
\name{fnch_render_pdf_document}
\alias{fnch_render_pdf_document}
\title{Render pdf documents}
\usage{
fnch_render_pdf_document(..., template_path, metadata)
}
\arguments{
\item{template_path}{Path of the latex template}

\item{metadata}{Vector with pandoc metadata}
}
\description{
Render pdf documents
}
\keyword{internal}
