% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury.R
\name{fnch_role_order}
\alias{fnch_role_order}
\title{Change the judge role to a string "number" for sorting}
\usage{
fnch_role_order(role)
}
\arguments{
\item{role}{Judge role}
}
\value{
A string
}
\description{
Change the judge role to a string "number" for sorting
}
\keyword{internal}
