% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_rulebook.R
\name{fnch_rulebook}
\alias{fnch_rulebook}
\title{Standard pdf document}
\usage{
fnch_rulebook(..., metadata = c())
}
\arguments{
\item{metadata}{Additional pandoc metadata}
}
\description{
Standard pdf document
}
