% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_sum_judges}
\alias{fnch_sum_judges}
\title{FSSE Accounting: Sum up judges}
\usage{
fnch_sum_judges(df)
}
\arguments{
\item{df}{A judge table}
}
\value{
A tibble
}
\description{
FSSE Accounting: Sum up judges
}
