% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_sum_old_judges}
\alias{fnch_sum_old_judges}
\title{FSSE Accounting: Sum up judges (old system)}
\usage{
fnch_sum_old_judges(df, tarif = 30)
}
\arguments{
\item{df}{A judges table (old system)}
}
\value{
A tibble
}
\description{
FSSE Accounting: Sum up judges (old system)
}
