% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_sum_presidency}
\alias{fnch_sum_presidency}
\title{FSSE Accounting: Sum up presidency}
\usage{
fnch_sum_presidency(df)
}
\arguments{
\item{df}{A presidency table}
}
\value{
A tibble
}
\description{
FSSE Accounting: Sum up presidency
}
