% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avsh_dressage.R
\name{get_avsh_championship_b_classes}
\alias{get_avsh_championship_b_classes}
\title{Get AVSH B Dressage classes}
\usage{
get_avsh_championship_b_classes()
}
\value{
A vector of strings
}
\description{
Get AVSH B Dressage classes
}
