% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avsh_dressage.R
\name{get_avsh_championship_m_classes}
\alias{get_avsh_championship_m_classes}
\title{Get AVSH M Dressage classes}
\usage{
get_avsh_championship_m_classes()
}
\value{
A vector of strings
}
\description{
Get AVSH M Dressage classes
}
