% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avsh_dressage.R
\name{get_avsh_championship_r_classes}
\alias{get_avsh_championship_r_classes}
\title{Get AVSH R Dressage classes}
\usage{
get_avsh_championship_r_classes()
}
\value{
A vector of strings
}
\description{
Get AVSH R Dressage classes
}
