% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avsh_dressage.R
\name{get_avsh_championship_ranking}
\alias{get_avsh_championship_ranking}
\title{Get AVSH Championship ranking list}
\usage{
get_avsh_championship_ranking(df, res, lic, ep_selection, kur = 0)
}
\arguments{
\item{df}{A results dataframe}

\item{res}{Number of results to consider}

\item{lic}{Licenses to consider (vector)}

\item{ep_selection}{Categories to consider}

\item{kur}{Number of Kürs to consider}
}
\value{
A cleaned up results dataframe
}
\description{
Get AVSH Championship ranking list
}
