% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avsh_dressage.R
\name{get_avsh_championship_s_classes}
\alias{get_avsh_championship_s_classes}
\title{Get AVSH S Dressage classes}
\usage{
get_avsh_championship_s_classes()
}
\value{
A vector of strings
}
\description{
Get AVSH S Dressage classes
}
