% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{get_clean_calendar}
\alias{get_clean_calendar}
\title{Cleanup calendar}
\usage{
get_clean_calendar(year)
}
\arguments{
\item{year}{An integer}
}
\value{
A tibble with the calendar
}
\description{
Cleanup calendar
}
