% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{get_events_fnch_login}
\alias{get_events_fnch_login}
\title{Get login for events.fnch.ch}
\usage{
get_events_fnch_login()
}
\value{
A list with id and token
}
\description{
Get login for events.fnch.ch
}
