% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fer_dressage.R
\name{get_fer_championship_events_list}
\alias{get_fer_championship_events_list}
\title{Wrapper around 'get_fnch_events' for FER events}
\usage{
get_fer_championship_events_list(startdate, enddate, ignore = c())
}
\arguments{
\item{startdate}{A stardate}

\item{enddate}{An enddate}
}
\value{
A list of events
}
\description{
Wrapper around 'get_fnch_events' for FER events
}
