% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fer_dressage.R
\name{get_fer_championship_n_classes}
\alias{get_fer_championship_n_classes}
\title{Get FER U21 Dressage classes}
\usage{
get_fer_championship_n_classes()
}
\value{
A vector of strings
}
\description{
Get FER U21 Dressage classes
}
