% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_dressage.R
\name{get_fer_championship_ranking}
\alias{get_fer_championship_ranking}
\title{Title}
\usage{
get_fer_championship_ranking(df, res, lic, ep_selection, kur = 0)
}
\arguments{
\item{df}{A results dataframe}

\item{res}{Number of results to consider}

\item{lic}{Licenses to consider (vector)}

\item{ep_selection}{Categories to consider}

\item{kur}{Number of Kürs to consider}
}
\value{
A cleaned up results dataframe
}
\description{
Title
}
