% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fer_dressage.R
\name{get_fer_championship_s_classes}
\alias{get_fer_championship_s_classes}
\title{Get FER S Dressage classes}
\usage{
get_fer_championship_s_classes()
}
\value{
A vector of strings
}
\description{
Get FER S Dressage classes
}
