% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fer_dressage.R
\name{get_fer_championship_swiss_r_classes}
\alias{get_fer_championship_swiss_r_classes}
\title{Get FER Swiss R Dressage classes}
\usage{
get_fer_championship_swiss_r_classes()
}
\value{
A vector of strings
}
\description{
Get FER Swiss R Dressage classes
}
