% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fer_dressage.R
\name{get_fer_championship_swiss_r_ranking}
\alias{get_fer_championship_swiss_r_ranking}
\title{Get FER Swiss R Championship ranking list}
\usage{
get_fer_championship_swiss_r_ranking(
  df,
  ep_selection,
  res = 4,
  lic = "DR",
  nb_sel = 1
)
}
\arguments{
\item{df}{A results dataframe}

\item{ep_selection}{Categories to consider}

\item{res}{Number of results to consider}

\item{lic}{Licenses to consider (vector)}

\item{kur}{Number of Kürs to consider}
}
\value{
A cleaned up results dataframe
}
\description{
Get FER Swiss R Championship ranking list
}
