% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_results.R
\name{get_fnch_class_results}
\alias{get_fnch_class_results}
\title{Get results for a specific class and event}
\usage{
get_fnch_class_results(id, eventid)
}
\arguments{
\item{id}{ID of the class}

\item{eventid}{ID of the event}
}
\value{
A dataframe with the results of the specified class
}
\description{
Get results for a specific class and event
}
