% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_events.R
\name{get_fnch_event_classes}
\alias{get_fnch_event_classes}
\title{Get all classes of an event}
\usage{
get_fnch_event_classes(eventid)
}
\arguments{
\item{eventid}{An event id}
}
\value{
List of classes
}
\description{
Get all classes of an event
}
