% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_events.R
\name{get_fnch_event_startlists}
\alias{get_fnch_event_startlists}
\title{Get all startlists of an event}
\usage{
get_fnch_event_startlists(eventid)
}
\arguments{
\item{eventid}{An event id}
}
\value{
A tibble
}
\description{
Get all startlists of an event
}
