% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_events.R
\name{get_fnch_events}
\alias{get_fnch_events}
\title{Get FNCH events from calendar}
\usage{
get_fnch_events(startdate, enddate, disziplin = "",
  regionalverband = c(), eventtyp = c(), typ = "")
}
\arguments{
\item{startdate}{A stardate}

\item{enddate}{An enddate}

\item{disziplin}{A discipline string}

\item{regionalverband}{A regional code}
}
\value{
List of events
}
\description{
Get FNCH events from calendar
}
