% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_horses.R
\name{get_fnch_horse_dressage_results}
\alias{get_fnch_horse_dressage_results}
\title{Get dressage results for horseid}
\usage{
get_fnch_horse_dressage_results(horseid)
}
\arguments{
\item{horseid}{A passport id for a horse}
}
\value{
A dataframe of dressage results
}
\description{
Get dressage results for horseid
}
