% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_horses.R
\name{get_fnch_horse_driving_results}
\alias{get_fnch_horse_driving_results}
\title{Get driving results for horseid}
\usage{
get_fnch_horse_driving_results(horseid)
}
\arguments{
\item{horseid}{A passport id for a horse}
}
\value{
A dataframe of driving results
}
\description{
Get driving results for horseid
}
