% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_horses.R
\name{get_fnch_horse_eventing_results}
\alias{get_fnch_horse_eventing_results}
\title{Get eventing results for horseid}
\usage{
get_fnch_horse_eventing_results(horseid)
}
\arguments{
\item{horseid}{A passport id for a horse}
}
\value{
A dataframe of eventing results
}
\description{
Get eventing results for horseid
}
