% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_horses.R
\name{get_fnch_horse_gwp}
\alias{get_fnch_horse_gwp}
\title{Get gwp infos for horseid}
\usage{
get_fnch_horse_gwp(horseid)
}
\arguments{
\item{horseid}{A passport id for a horse}
}
\value{
A dataframe of gwp data
}
\description{
Get gwp infos for horseid
}
