% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_horses.R
\name{get_fnch_horse_infos}
\alias{get_fnch_horse_infos}
\title{Get general infos for horseid}
\usage{
get_fnch_horse_infos(horseid)
}
\arguments{
\item{horseid}{A passport id for a horse}
}
\value{
A list of data
}
\description{
Get general infos for horseid
}
