% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_horses.R
\name{get_fnch_horse_jumping_gwp}
\alias{get_fnch_horse_jumping_gwp}
\title{Get Horse Jumping GWP}
\usage{
get_fnch_horse_jumping_gwp(
  horseid,
  enddate,
  startdate = floor_date(lubridate::today(), "year") - years(1)
)
}
\arguments{
\item{horseid}{A passport id for a horse}

\item{enddate}{The end date for the GWP period}

\item{startdate}{The start date for the GWP. Defaults to last year.}
}
\value{
An integer amount of points.
}
\description{
Get Horse Jumping GWP
}
