% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_horses.R
\name{get_fnch_horse_jumping_results}
\alias{get_fnch_horse_jumping_results}
\title{Get jumping results for horseid}
\usage{
get_fnch_horse_jumping_results(horseid)
}
\arguments{
\item{horseid}{A passport id for a horse}
}
\value{
A dataframe of jumping results
}
\description{
Get jumping results for horseid
}
