% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_horses.R
\name{get_fnch_horse_results}
\alias{get_fnch_horse_results}
\title{Get results for horseid}
\usage{
get_fnch_horse_results(horseid, discipline)
}
\arguments{
\item{horseid}{A passport id for a horse}

\item{discipline}{The discipline}
}
\value{
A dataframe of results
}
\description{
Get results for horseid
}
