% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_databases.R, R/fnch_horses.R
\name{get_fnch_horses}
\alias{get_fnch_horses}
\title{Return the horses' database}
\usage{
get_fnch_horses()

get_fnch_horses()
}
\value{
A dataframe with all FNCH horses

A dataframe
}
\description{
Return the horses' database

Get general horse table
}
