% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_logbook.R
\name{get_fnch_logbook_entries}
\alias{get_fnch_logbook_entries}
\title{Get Logbook entries for a rider id}
\usage{
get_fnch_logbook_entries(rider_id)
}
\arguments{
\item{rider_id}{Rider licence id}
}
\value{
A list of events
}
\description{
Get Logbook entries for a rider id
}
