% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_logbook.R
\name{get_fnch_logbook_events}
\alias{get_fnch_logbook_events}
\title{Get Events for speaker selection}
\usage{
get_fnch_logbook_events()
}
\value{
A tibble
}
\description{
Get Events for speaker selection
}
