% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_officials.R
\name{get_fnch_officials_full}
\alias{get_fnch_officials_full}
\title{Get officials full (with addresses)}
\usage{
get_fnch_officials_full()
}
\value{
A dataframe
}
\description{
Get officials full (with addresses)
}
