% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_databases.R
\name{get_fnch_rider_address}
\alias{get_fnch_rider_address}
\title{Get rider address}
\usage{
get_fnch_rider_address(license)
}
\arguments{
\item{license}{A rider license id}
}
\value{
A string with rider name and address
}
\description{
Get rider address
}
