% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_databases.R
\name{get_fnch_rider_adress_table}
\alias{get_fnch_rider_adress_table}
\title{Get rider address and return it as a dataframe}
\usage{
get_fnch_rider_adress_table(license)
}
\arguments{
\item{license}{A rider license id}
}
\value{
A dataframe with Licence, FirstName, LastName, Address, ZIP and Place
}
\description{
Get rider address and return it as a dataframe
}
