% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_riders.R
\name{get_fnch_rider_driving_results}
\alias{get_fnch_rider_driving_results}
\title{Get driving results for riderid}
\usage{
get_fnch_rider_driving_results(riderid)
}
\arguments{
\item{riderid}{A licence id for a rider}
}
\value{
A dataframe of driving results
}
\description{
Get driving results for riderid
}
