% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_databases.R
\name{get_fnch_rider_email}
\alias{get_fnch_rider_email}
\title{Get rider email}
\usage{
get_fnch_rider_email(license)
}
\arguments{
\item{license}{A rider license id}
}
\value{
A string with rider name and email
}
\description{
Get rider email
}
