% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_riders.R
\name{get_fnch_rider_eventing_results}
\alias{get_fnch_rider_eventing_results}
\title{Get eventing results for riderid}
\usage{
get_fnch_rider_eventing_results(riderid)
}
\arguments{
\item{riderid}{A licence id for a rider}
}
\value{
A dataframe of eventing results
}
\description{
Get eventing results for riderid
}
