% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_riders.R
\name{get_fnch_rider_gwp}
\alias{get_fnch_rider_gwp}
\title{Get gwp infos for riderid}
\usage{
get_fnch_rider_gwp(riderid)
}
\arguments{
\item{riderid}{A licence id for a rider}
}
\value{
A dataframe of gwp data
}
\description{
Get gwp infos for riderid
}
