% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_riders.R
\name{get_fnch_rider_infos}
\alias{get_fnch_rider_infos}
\title{Get general infos for riderid}
\usage{
get_fnch_rider_infos(riderid)
}
\arguments{
\item{riderid}{A licence id for a rider}
}
\value{
A list of rider infos
}
\description{
Get general infos for riderid
}
