% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_databases.R
\name{get_fnch_rider_license}
\alias{get_fnch_rider_license}
\title{Get rider licence infos}
\usage{
get_fnch_rider_license(license)
}
\arguments{
\item{riderid}{A rider license id}
}
\value{
A string with rider name and licence infos
}
\description{
Get rider licence infos
}
