% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_databases.R
\name{get_fnch_rider_mobile}
\alias{get_fnch_rider_mobile}
\title{Get rider phone}
\usage{
get_fnch_rider_mobile(license)
}
\arguments{
\item{license}{A rider license id}
}
\value{
A string with rider name and mobile phone
}
\description{
Get rider phone
}
