% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_riders.R
\name{get_fnch_rider_results}
\alias{get_fnch_rider_results}
\title{Get results for riderid}
\usage{
get_fnch_rider_results(riderid, discipline)
}
\arguments{
\item{riderid}{A licence id for a rider}

\item{discipline}{The discipline}
}
\value{
A dataframe of results
}
\description{
Get results for riderid
}
