% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_class_cat}
\alias{get_fnch_sp_class_cat}
\title{Get the class selection}
\usage{
get_fnch_sp_class_cat(code)
}
\arguments{
\item{code}{}
}
\value{
A string vector
}
\description{
Get the class selection
}
