% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_class_cat_cc}
\alias{get_fnch_sp_class_cat_cc}
\title{Return list of possible eventing category choices}
\usage{
get_fnch_sp_class_cat_cc()
}
\value{
A string vector
}
\description{
Return list of possible eventing category choices
}
