% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_class_cat_dr}
\alias{get_fnch_sp_class_cat_dr}
\title{Return list of possible dressage category choices}
\usage{
get_fnch_sp_class_cat_dr()
}
\value{
A string vector
}
\description{
Return list of possible dressage category choices
}
