% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_class_cat_fa}
\alias{get_fnch_sp_class_cat_fa}
\title{Return list of possible driving category choices}
\usage{
get_fnch_sp_class_cat_fa()
}
\value{
A string vector
}
\description{
Return list of possible driving category choices
}
