% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_class_min_dr}
\alias{get_fnch_sp_class_min_dr}
\title{Function to get the dressage classes}
\usage{
get_fnch_sp_class_min_dr(class_min)
}
\arguments{
\item{class_min}{The minimal class}
}
\description{
Function to get the dressage classes
}
