% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_clean_results}
\alias{get_fnch_sp_clean_results}
\title{Format the results and return the clean result table}
\usage{
get_fnch_sp_clean_results(results, discipline)
}
\arguments{
\item{results}{The results}

\item{discipline}{The discipline}
}
\description{
Format the results and return the clean result table
}
