% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_details}
\alias{get_fnch_sp_details}
\title{Generate the details for each rider}
\usage{
get_fnch_sp_details(index, startlist, results, titles, discipline)
}
\arguments{
\item{index}{The row index}

\item{startlist}{The startlist}

\item{results}{The results}

\item{titles}{The titles}
}
\description{
Generate the details for each rider
}
