% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_details_titles}
\alias{get_fnch_sp_details_titles}
\title{Return table with rider titles}
\usage{
get_fnch_sp_details_titles(rider_id, titles)
}
\arguments{
\item{rider_id}{The rider id}

\item{titles}{The title table}
}
\description{
Return table with rider titles
}
