% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_dr_programme_fr}
\alias{get_fnch_sp_dr_programme_fr}
\title{Return french dressage programmes}
\usage{
get_fnch_sp_dr_programme_fr(programme)
}
\value{
A string vector
}
\description{
Return french dressage programmes
}
\keyword{internal}
