% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_events}
\alias{get_fnch_sp_events}
\title{Get events}
\usage{
get_fnch_sp_events(discipline)
}
\arguments{
\item{discipline}{Discipline code}
}
\value{
A list of events
}
\description{
Get events
}
