% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_events_dr}
\alias{get_fnch_sp_events_dr}
\title{Get Events for speaker dressage selection}
\usage{
get_fnch_sp_events_dr()
}
\value{
A tibble
}
\description{
Get Events for speaker dressage selection
}
