% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_fixed_results}
\alias{get_fnch_sp_fixed_results}
\title{Make a pferd_id for eventing results}
\usage{
get_fnch_sp_fixed_results(results, discipline)
}
\arguments{
\item{results}{The results}

\item{discipline}{The discipline code}
}
\value{
A dataframe with horse_id
}
\description{
Make a pferd_id for eventing results
}
