% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_horse_details}
\alias{get_fnch_sp_horse_details}
\title{Get details for all horseid}
\usage{
get_fnch_sp_horse_details(horseids)
}
\arguments{
\item{horseids}{A vector of horseids}
}
\value{
A tibble
}
\description{
Get details for all horseid
}
