% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_horse_origins}
\alias{get_fnch_sp_horse_origins}
\title{Get the hors origins table}
\usage{
get_fnch_sp_horse_origins(starter, pdf = FALSE)
}
\arguments{
\item{starter}{The starter data}
}
\value{
A dataframe with horse origins
}
\description{
Get the hors origins table
}
