% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_startlist}
\alias{get_fnch_sp_startlist}
\title{Make startlist table}
\usage{
get_fnch_sp_startlist(startlist_data, titles)
}
\arguments{
\item{eventid}{Event id}

\item{classid}{Classid}

\item{nb_years}{Number of years to go back}

\item{nb_ranks}{Number of ranks to consider}
}
\value{
A reactable object
}
\description{
Make startlist table
}
