% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_startlist_data}
\alias{get_fnch_sp_startlist_data}
\title{Get startlist data}
\usage{
get_fnch_sp_startlist_data(eventid, classid, nb_years, nb_ranks, class_min)
}
\arguments{
\item{eventid}{Event ID}

\item{classid}{Class ID}

\item{nb_years}{Number of years}

\item{nb_ranks}{Number of rankings}

\item{class_min}{Mininum height for class}
}
\value{
A tibble
}
\description{
Get startlist data
}
