% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_startlist_head}
\alias{get_fnch_sp_startlist_head}
\title{Select startlist head}
\usage{
get_fnch_sp_startlist_head(startlist, discipline)
}
\arguments{
\item{startlist}{The startlist}

\item{discipline}{The discipline code}
}
\value{
A daframe
}
\description{
Select startlist head
}
