% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{get_fnch_sp_titles}
\alias{get_fnch_sp_titles}
\title{Retrieve title table from Notion}
\usage{
get_fnch_sp_titles(titles_min)
}
\arguments{
\item{titles_min}{Min year for titles}
}
\value{
A tibble
}
\description{
Retrieve title table from Notion
}
