% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_events.R
\name{get_fnch_startlist}
\alias{get_fnch_startlist}
\title{Get a specific startlist}
\usage{
get_fnch_startlist(eventid, classid)
}
\arguments{
\item{eventid}{An event id}

\item{classid}{A class id}
}
\value{
A tibble
}
\description{
Get a specific startlist
}
