% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{get_my_fnch_authenticy_token}
\alias{get_my_fnch_authenticy_token}
\title{Get authenticity token for my.fnch.ch}
\usage{
get_my_fnch_authenticy_token()
}
\value{
A string with authenticity token for login function
}
\description{
Get authenticity token for my.fnch.ch
}
