% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{get_my_fnch_credentials}
\alias{get_my_fnch_credentials}
\title{Get credentials for my.fnch.ch}
\usage{
get_my_fnch_credentials()
}
\value{
A list with id and token
}
\description{
Get credentials for my.fnch.ch
}
