% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{get_my_fnch_dokumente}
\alias{get_my_fnch_dokumente}
\title{Get document info for event}
\usage{
get_my_fnch_dokumente(id)
}
\arguments{
\item{id}{Event id}
}
\value{
A dataframe
}
\description{
Get document info for event
}
