% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{get_my_fnch_event_details}
\alias{get_my_fnch_event_details}
\title{Get all details for event}
\usage{
get_my_fnch_event_details(id)
}
\arguments{
\item{id}{Event id}
}
\value{
A list
}
\description{
Get all details for event
}
