% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{get_my_fnch_events}
\alias{get_my_fnch_events}
\title{Get event table from my.fnch.ch}
\usage{
get_my_fnch_events(year)
}
\arguments{
\item{year}{An integer}
}
\value{
A dataframe
}
\description{
Get event table from my.fnch.ch
}
