% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{get_my_fnch_events_excel}
\alias{get_my_fnch_events_excel}
\title{Get internal my.fnch.ch calendar}
\usage{
get_my_fnch_events_excel(year)
}
\arguments{
\item{year}{The year of the calendar}
}
\value{
A tibble
}
\description{
Get internal my.fnch.ch calendar
}
