% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_my_portal.R
\name{get_my_fnch_events_summary}
\alias{get_my_fnch_events_summary}
\title{Get summary table of number of starters per event organizer}
\usage{
get_my_fnch_events_summary(year)
}
\arguments{
\item{year}{An integer}
}
\value{
A tibble
}
\description{
Get summary table of number of starters per event organizer
}
