% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{get_my_fnch_rider_details}
\alias{get_my_fnch_rider_details}
\title{Get all details for a person}
\usage{
get_my_fnch_rider_details(id)
}
\arguments{
\item{id}{Person id}
}
\value{
A list
}
\description{
Get all details for a person
}
