% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_rosson.R
\name{get_rosson_latest_version}
\alias{get_rosson_latest_version}
\title{Get the latest Rosson version string}
\usage{
get_rosson_latest_version()
}
\value{
A string
}
\description{
Get the latest Rosson version string
}
