% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_markdown.R
\name{inherit_pdf_document}
\alias{inherit_pdf_document}
\title{Rmarkdown pdf document}
\usage{
inherit_pdf_document(...)
}
\arguments{
\item{...}{}
}
\description{
Call rmarkdown::pdf_document and mark the return
value as inheriting pdf_document
}
\keyword{internal}
