% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_knit.R
\name{knit_all}
\alias{knit_all}
\title{Knit all Rmd files in the Project}
\usage{
knit_all(path = here::here())
}
\arguments{
\item{path}{The path to start at}
}
\description{
Knit all Rmd files in the Project
}
