% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_knit.R
\name{knit_cleanup}
\alias{knit_cleanup}
\title{Remove all temporary files}
\usage{
knit_cleanup(input)
}
\arguments{
\item{path}{A folder to check for files}
}
\description{
Remove all temporary files
}
\keyword{internal}
