% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_markdown.R
\name{knit_metadata}
\alias{knit_metadata}
\title{Generate pandoc metadata}
\usage{
knit_metadata(path)
}
\arguments{
\item{path}{Path of template files}
}
\value{
A vector with pandoc metadata arguments
}
\description{
Generate pandoc metadata
}
\keyword{internal}
