% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_knit.R
\name{knit_subfile}
\alias{knit_subfile}
\title{Knit subfiles}
\usage{
knit_subfile(input, output, key, value)
}
\arguments{
\item{input}{The input file}

\item{keyword}{The subfile head as a named vector}
}
\description{
Knit subfiles
}
\keyword{internal}
