% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_dressage.R
\name{map_category_arguments}
\alias{map_category_arguments}
\title{Function to build arguments for category list}
\usage{
map_category_arguments(i)
}
\arguments{
\item{i}{Number of arguments to generate for category}
}
\description{
Function to build arguments for category list
}
