% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_classes.R
\name{map_class_results}
\alias{map_class_results}
\title{Function to map class results}
\usage{
map_class_results(class_obj)
}
\arguments{
\item{class_obj}{A class object}
}
\value{
A dataframe with results of class, function adds some additional info
}
\description{
Function to map class results
}
