% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_results.R
\name{map_fnch_class_results}
\alias{map_fnch_class_results}
\title{Function to map class results}
\usage{
map_fnch_class_results(class_table)
}
\arguments{
\item{class_obj}{A dataframe with class ids}
}
\value{
A dataframe with class results
}
\description{
Function to map class results
}
