% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_events.R
\name{map_fnch_event_classes}
\alias{map_fnch_event_classes}
\title{Map classes to event table}
\usage{
map_fnch_event_classes(event_table)
}
\arguments{
\item{event_table}{A dataframe from `get_fnch_events`}
}
\value{
A dataframe
}
\description{
Map classes to event table
}
