% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_riders.R
\name{map_fnch_riders_ldap}
\alias{map_fnch_riders_ldap}
\title{Format the official data for LDAP}
\usage{
map_fnch_riders_ldap(...)
}
\arguments{
\item{...}{Data from full official database}
}
\value{
A list with the parameters for the LDAP CSV
}
\description{
Format the official data for LDAP
}
\keyword{internal}
