% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_dressage.R
\name{map_ranking_arguments}
\alias{map_ranking_arguments}
\title{Function to build arguments for ranking list}
\usage{
map_ranking_arguments(i)
}
\arguments{
\item{i}{Number of arguments}
}
\value{
A list of arguments
}
\description{
Function to build arguments for ranking list
}
