% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.R
\name{pr_fnch_calendar}
\alias{pr_fnch_calendar}
\title{Plumber FNCH calendar}
\usage{
pr_fnch_calendar(year, federation)
}
\arguments{
\item{year}{An integer: the calendar year}

\item{federation}{A string: the regional federation code. FER returns a combined set.}
}
\value{
A string: ICS calendar
}
\description{
Plumber FNCH calendar
}
