% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.R
\name{pr_fnch_full_calendar}
\alias{pr_fnch_full_calendar}
\title{Plumber FNCH full calendar}
\usage{
pr_fnch_full_calendar(federation)
}
\arguments{
\item{federation}{A string: the regional federation code. FER returns a combined set.}
}
\value{
A string: ICS calendar
}
\description{
Plumber FNCH full calendar
}
