% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.R
\name{pr_fnch_xls_calendar}
\alias{pr_fnch_xls_calendar}
\title{Plumber FNCH excel calendar}
\usage{
pr_fnch_xls_calendar(year)
}
\arguments{
\item{year}{An integer: the year for the desired calendar}
}
\value{
A binary excel file
}
\description{
Plumber FNCH excel calendar
}
