% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_dressage.R
\name{rename_ranking_columns}
\alias{rename_ranking_columns}
\title{Rename the columns of the fer championship ranking}
\usage{
rename_ranking_columns(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
A dataframe
}
\description{
Rename the columns of the fer championship ranking
}
