% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_speaker.R
\name{render_fnch_sp_startlist_pdf}
\alias{render_fnch_sp_startlist_pdf}
\title{Make PDF startlist}
\usage{
render_fnch_sp_startlist_pdf(startlist_data, titles)
}
\arguments{
\item{startlist_data}{The startlist data}

\item{titles}{The titles data}
}
\value{
A path to the pdf
}
\description{
Make PDF startlist
}
\seealso{
[get_fnch_startlist_data()]
}
