% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_ranking.R
\name{return_discipline_id}
\alias{return_discipline_id}
\title{Change discipline abreviation to code}
\usage{
return_discipline_id(x)
}
\arguments{
\item{x}{A string, either "CS", "CD" or "CC"}
}
\value{
An integer or vector of integers
}
\description{
Change discipline abreviation to code
}
