% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_events.R
\name{return_event_typ_id}
\alias{return_event_typ_id}
\title{Change event types to codes}
\usage{
return_event_typ_id(x)
}
\arguments{
\item{x}{A string or vector of strings}
}
\value{
An id or vector of ids
}
\description{
Change event types to codes
}
