% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fer_dressage.R
\name{return_fer_category_arguments}
\alias{return_fer_category_arguments}
\title{Generate category arguments}
\usage{
return_fer_category_arguments(x)
}
\arguments{
\item{x}{Number of arguments to generate for category}
}
\value{
List of category arguments
}
\description{
Generate category arguments
}
