% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_dressage.R
\name{return_fer_ranking_arguments}
\alias{return_fer_ranking_arguments}
\title{Generate arguments}
\usage{
return_fer_ranking_arguments(x)
}
\arguments{
\item{x}{Number of arguments for ranking list to generate}
}
\value{
List of arguments
}
\description{
Generate arguments
}
