% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_events.R
\name{return_regional_id}
\alias{return_regional_id}
\title{Change regional federation abreviation to codes}
\usage{
return_regional_id(x)
}
\arguments{
\item{x}{A string or vector of strings}
}
\value{
An id or vector of ids
}
\description{
Change regional federation abreviation to codes
}
