% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumber.R
\name{run_api}
\alias{run_api}
\title{Run Plumber API}
\usage{
run_api(port = 3806, host = "0.0.0.0", ...)
}
\arguments{
\item{port}{An integer: the port to run the API on}

\item{host}{A string: the host to bind to}

\item{...}{Options passed to \code{plumber::plumb()$run()}}
}
\value{
A running Plumber API
}
\description{
Run Plumber API
}
