% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_backend.R
\name{sum_my_fnch_taxes}
\alias{sum_my_fnch_taxes}
\title{Sum the taxes pulled from the account pdf}
\usage{
sum_my_fnch_taxes(pdf_text)
}
\arguments{
\item{pdf_text}{A string}
}
\value{
An integer
}
\description{
Sum the taxes pulled from the account pdf
}
