% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_knit.R
\name{tex_global_path}
\alias{tex_global_path}
\title{Return full path to geovizr ressources}
\usage{
tex_global_path()
}
\value{
A string with the full path
}
\description{
Return full path to geovizr ressources
}
\keyword{internal}
