% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{write_fnch_fer_calendar}
\alias{write_fnch_fer_calendar}
\title{Write FER calendar}
\usage{
write_fnch_fer_calendar(year, path)
}
\arguments{
\item{year}{An integer}

\item{path}{A path}
}
\description{
Write FER calendar
}
