% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{write_fnch_ics_calendar}
\alias{write_fnch_ics_calendar}
\title{Write ICS calendar}
\usage{
write_fnch_ics_calendar(year, federation, path)
}
\arguments{
\item{year}{An integer for the desired calendar year}

\item{federation}{A string vector with the desired federations}

\item{path}{A path to write the ICS file to}
}
\description{
Write ICS calendar
}
