% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_officials.R
\name{write_fnch_officials_csv}
\alias{write_fnch_officials_csv}
\title{Write CSV official data file for LDAP}
\usage{
write_fnch_officials_csv()
}
\value{
Nothing (writes csv file)
}
\description{
Write CSV official data file for LDAP
}
