% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_riders.R
\name{write_fnch_riders_csv}
\alias{write_fnch_riders_csv}
\title{Write CSV rider data file for LDAP}
\usage{
write_fnch_riders_csv()
}
\value{
Nothing (writes csv file)
}
\description{
Write CSV rider data file for LDAP
}
