% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{write_fnch_week_calendar}
\alias{write_fnch_week_calendar}
\title{Write calendar}
\usage{
write_fnch_week_calendar(
  path,
  year,
  federations = c("AEN", "ASCJ", "AVSH", "FFSE", "FGE", "SCV", "ZKV")
)
}
\arguments{
\item{path}{A path}

\item{year}{An integer}

\item{federations}{A vector of federation abreviations}
}
\description{
Write calendar
}
