% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fer_dressage.R
\name{add_category_r}
\alias{add_category_r}
\title{Determine level for R championship}
\usage{
add_category_r(df, nb_m, max_pt)
}
\arguments{
\item{df}{A results dataframe}

\item{nb_m}{Max number of M results}

\item{max_pt}{Max number of points for L level}
}
\value{
A results dataframe
}
\description{
Determine level for R championship
}
