% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aen_jumping.R
\name{aen_get_championship_data}
\alias{aen_get_championship_data}
\title{Get data for championship qualification analysis}
\usage{
aen_get_championship_data(riderid, horseid, last_championship)
}
\arguments{
\item{riderid}{A licence id for a rider}

\item{horseid}{A passport id for a horse}

\item{last_championship}{The date of the last championship}
}
\value{
A list of data points for use in qualification functions
}
\description{
Get data for championship qualification analysis
}
