% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_jury_admin.R
\name{fnch_add_candidate}
\alias{fnch_add_candidate}
\title{FSSE Accounting: Add row to judges table}
\usage{
fnch_add_candidate(df, name, days, classes)
}
\arguments{
\item{df}{A judge table}

\item{name}{Name of judge}

\item{days}{Number of days judged}

\item{classes}{Number of classes}
}
\value{
A tibble
}
\description{
FSSE Accounting: Add row to judges table
}
