% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{get_fnch_ics_calendar}
\alias{get_fnch_ics_calendar}
\title{Get ICS calendar}
\usage{
get_fnch_ics_calendar(year, federation)
}
\arguments{
\item{year}{An integer for the desired calendar year}

\item{federation}{A string vector with the desired federations}
}
\value{
An ICS calendar in string format
}
\description{
Get ICS calendar
}
