% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_ranking.R
\name{get_fnch_ranking}
\alias{get_fnch_ranking}
\title{Get FNCH ranking list}
\usage{
get_fnch_ranking(start, end, discipline = "CS", canton = "")
}
\arguments{
\item{start}{A date for the start of the period}

\item{end}{A date for the end of the period}

\item{discipline}{A string, either "CS", "CD" or "CC"}

\item{canton}{A string for the canton}
}
\value{
A dataframe with the rankings
}
\description{
Get FNCH ranking list
}
