% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_classes.R
\name{get_fnch_result_frame}
\alias{get_fnch_result_frame}
\title{Get results for all events}
\usage{
get_fnch_result_frame(event_obj, classfilter = "modus_code == \\"O\\"")
}
\arguments{
\item{event_obj}{An event object list}
}
\value{
A dataframe with results for all events
}
\description{
Get results for all events
}
