% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{write_fnch_collision_calendar}
\alias{write_fnch_collision_calendar}
\title{Write collision calendar for federation}
\usage{
write_fnch_collision_calendar(path, year, federation)
}
\arguments{
\item{path}{A path}

\item{year}{An integer}

\item{federation}{A string}
}
\description{
Write collision calendar for federation
}
