% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnch_calendar.R
\name{write_fnch_ics_calendar}
\alias{write_fnch_ics_calendar}
\title{Write ICS calendar}
\usage{
write_fnch_ics_calendar(start, end, federation, path)
}
\arguments{
\item{start}{A string for the start date (format yyyy-mm-dd)}

\item{end}{A string for the end date (format yyyy-mm-dd)}

\item{federation}{A string vector with the desired federations}

\item{path}{A path to write the ICS file to}
}
\description{
Write ICS calendar
}
